feat(settings): 设置界面重构 — 分节化 / 精简 / Beta 渠道 / 麦克风 / 市场登录
将 2200 行的 Settings.tsx 单体拆为 pages/settings/ 下的按节文件（通用 / 服务 /
隐私 / 高级 / 关于），并对设置界面做整体精简与功能调整。

界面精简
- 移除各设置项冗长 / 多行描述（在 180px 标签列普遍折成多行），只留标签与控件
- 删除「个性化」独立 tab —— 仅剩字体大小一项、整页太空，并入「关于」
- 移除失效的「毛玻璃强度」滑块（其 CSS 变量无可见元素消费）

关于 / 更新
- 版本卡片图标右上方放「检查更新」按钮，固定查正式版（channel='stable'）
- 「加入 Beta 渠道」迁到「高级」页底部、单独成节；开启后出现查测试版的检查按钮
- 检查中按钮内转圈；结果（已是最新 / 失败）只在按钮内呈现，卡片高度恒定，
  不再因状态文字 reflow 抽搐
- Rust app_check_update_with_channel 新增可选 channel 参数，两个按钮各查各渠道；
  后台 AutoUpdateGate 仍按 prefs.update_channel 走

麦克风选择
- 由全屏弹窗改为 SelectLite 下拉（与语言选择一致的「官方框」）
- SelectLite 增加 trailing / onOpenChange；新增 MicrophoneSelect —— 选中项右侧
  打勾、勾左侧实时音量条

市场登录
- 抽出共用 GithubLoginModal 与 ui/Modal，扩展市场与风格市场登录统一走弹窗
- 修复风格市场登录弹窗「先小后大」尺寸跳动（各阶段固定 minHeight）

弹窗动画
- 去掉开启动画里的 blur 滤镜（卡顿 / 闪烁），改纯 opacity + transform
- 设置弹窗切 tab 内容轻微淡入；各处弹窗动画时长 / 缓动统一到 global.css

网络
- 新增 net.rs：进程级共享 reqwest 客户端（连接池复用）+ 传输层失败指数退避重试
