+++ openless-all/app/src-tauri/src/net.rs
@@ -1,0 +1,68 @@
+//! 共享 HTTP 客户端 + 带重试的请求发送。
+//!
+//! 背景：原先每个网络命令各自 `reqwest::Client::new()`，连接池互不复用 —— 一次
+//! 成功的 TLS 连接用完即弃，下一个命令又得重新握手。在握手不稳定的网络下（代理
+//! 分流等）首次握手经常被重置，用户得反复重试才能用。
+//!
+//! 这里提供两件东西：
+//! - `http()`：进程级共享客户端。一次握手成功后的连接进连接池，后续命令直接复用，
+//!   不再付握手成本。
+//! - `send_with_retry`：对「连接 / 超时 / 请求未送出」这三类传输层失败做指数退避
+//!   重试。这类失败发生在请求送达服务端之前，重试是幂等安全的。HTTP 4xx/5xx 不在
+//!   重试范围 —— 服务端已经应答，状态码交给调用方判断。
+
+use std::time::Duration;
+
+use once_cell::sync::Lazy;
+
+static HTTP: Lazy<reqwest::Client> = Lazy::new(|| {
+    reqwest::Client::builder()
+        // 握手单独限时：卡在握手上要尽快失败，好让 send_with_retry 立即重试。
+        .connect_timeout(Duration::from_secs(8))
+        // 连接池：一条握手成功的连接保留 90s 供后续命令复用。
+        .pool_idle_timeout(Duration::from_secs(90))
+        .pool_max_idle_per_host(8)
+        .tcp_keepalive(Duration::from_secs(30))
+        .user_agent(concat!("OpenLess/", env!("CARGO_PKG_VERSION")))
+        .build()
+        .unwrap_or_else(|_| reqwest::Client::new())
+});
+
+/// 进程级共享 HTTP 客户端。带连接池 —— 一次握手成功后的连接被后续请求复用。
+pub fn http() -> &'static reqwest::Client {
+    &HTTP
+}
+
+/// 单次请求最多尝试的次数。失败本身很快（握手重置 ~0.5s），10 次总耗时仍可控。
+const MAX_ATTEMPTS: u32 = 10;
+
+/// 发送请求，对传输层失败（连接重置 / 超时 / 请求未送出）做指数退避重试。
+///
+/// `make` 每次尝试都重新构造 `RequestBuilder`（`send()` 会消耗它）。只重试
+/// `is_connect()` / `is_timeout()` / `is_request()` —— 这三类失败时请求尚未送达
+/// 服务端，重试幂等安全。拿到任意 HTTP 响应（含 4xx/5xx）即返回，状态码由调用方
+/// 自行判断。
+pub async fn send_with_retry<F>(make: F) -> reqwest::Result<reqwest::Response>
+where
+    F: Fn() -> reqwest::RequestBuilder,
+{
+    let mut attempt: u32 = 0;
+    loop {
+        attempt += 1;
+        match make().send().await {
+            Ok(resp) => return Ok(resp),
+            Err(err) => {
+                let retryable = err.is_connect() || err.is_timeout() || err.is_request();
+                if !retryable || attempt >= MAX_ATTEMPTS {
+                    return Err(err);
+                }
+                // 150 / 300 / 600 / 900 / 900 … ms 退避。
+                let backoff = (150u64 * 2u64.pow((attempt - 1).min(3))).min(900);
+                log::warn!(
+                    "[net] transient failure (attempt {attempt}/{MAX_ATTEMPTS}), retry in {backoff}ms: {err}"
+                );
+                tokio::time::sleep(Duration::from_millis(backoff)).await;
+            }
+        }
+    }
+}
