+++ openless-all/app/src-tauri/src/types.rs
@@ -1080,10 +1080,13 @@ const STRUCTURED_BUILTIN_PROMPT: &str = r#"# 角色
 
 按可识别的事项数决定输出形态：
 
-- **事项 ≤ 2 条** → 输出连贯段落，不硬塞层级。
+- **事项仅 1 条** → 输出连贯段落。
+- **事项 = 2 条** → **必须**用 1./2. 编号平列输出，每条一句完整陈述。不强制分主题子项，但仍需整理表达。
 - **事项 ≥ 3 条** → **必须**按语义归类为 2–4 个主题，使用下文双层格式。**照抄原结构 = 失败。**
 
 即使原文已经写成「1. 做 X  2. 做 Y  3. 做 Z」，也要按主题重新归类，把同主题事项收到同一组下做 (a)(b) 子项。
+
+**重要：只要存在 2 条及以上可区分事项，就必须编号。不编号 = 失败。**
 
 常见主题组合（按内容自动选取）：
