+++ openless-all/app/src/i18n/en.ts
@@ -495,17 +495,31 @@ export const en: typeof zhCN = {
     kicker: 'SETTINGS',
     title: 'Settings',
     desc: 'Recording, providers, shortcuts, and permissions.',
-    sections: {
-      recording: 'Recording',
-      providers: 'Providers',
-      shortcuts: 'Shortcuts',
-      permissions: 'Permissions',
-      language: 'Language',
-      advanced: 'Advanced',
-      about: 'About',
+    dataStorage: {
+      title: 'Data storage',
+      desc: 'Conversation history and context kept on this device.',
+    },
+    debug: {
+      title: 'Debug tools',
+      desc: 'For troubleshooting recognition issues; off by default.',
+    },
+    marketplace: {
+      title: 'Marketplace',
+      desc: 'Upload identity for the style marketplace. Browse and install styles on the Styles page.',
+      github: {
+        signIn: 'Sign in with GitHub',
+        signedIn: 'Signed in with GitHub',
+        signedOut: 'Sign in to upload styles and like packs.',
+        signOut: 'Sign out',
+        starting: 'Starting sign-in…',
+        codeHint: 'Enter this code on the GitHub page that just opened:',
+        openGithub: 'Open GitHub',
+        waiting: 'GitHub opened — you’ll be signed in once you authorize…',
+        failed: 'Sign-in failed, please retry',
+      },
     },
     recording: {
-      title: 'Recording',
+      title: 'Recording & input',
       desc: 'Global recording hotkey and trigger mode.',
       hotkeyLabel: 'Recording hotkey',
       hotkeyDescAcc: 'Press to capture voice globally (requires Accessibility permission).',
@@ -765,8 +779,9 @@ export const en: typeof zhCN = {
       privacy: 'Privacy',
       privacyDesc: 'All data stays on this device. Cloud APIs do not retain recordings.',
       localFirst: 'Local-first',
+      linksTitle: 'Documentation',
       betaChannelLabel: 'Join Beta channel',
-      betaChannelDesc: 'Receive Beta updates when enabled. May be unstable — recommended only for early adopters.',
+      betaChannelDesc: 'Early access to new features',
       betaChannelFetching: 'Fetching the latest Beta…',
       betaChannelFetchBtn: 'Look up latest Beta',
       betaChannelLatestPrefix: 'Latest Beta:',
@@ -807,25 +822,16 @@ export const en: typeof zhCN = {
   },
   modal: {
     sections: {
-      account: 'Account',
-      settings: 'Settings',
+      general: 'General',
+      services: 'Services',
+      privacy: 'Privacy',
+      advanced: 'Advanced',
       personalize: 'Personalize',
       about: 'About',
       helpCenter: 'Help center',
       releaseNotes: 'Release notes',
     },
-    account: {
-      localUser: 'Local user',
-      localUserDesc: 'Not signed in · all data stays local',
-      loginSync: 'Sign in / Sync',
-      footer: 'Runs fully locally by default. Sign in to sync vocabulary and style presets across devices.',
-    },
     personalize: {
-      appearance: 'Appearance',
-      appearanceDesc: 'Follow system / Light / Dark',
-      appearanceSystem: 'Follow system',
-      appearanceLight: 'Light',
-      appearanceDark: 'Dark',
       font: 'Font size',
       fontDesc: 'Scale the entire UI font size — applies instantly.',
       fontSmall: 'Small',
@@ -833,10 +839,6 @@ export const en: typeof zhCN = {
       fontLarge: 'Large',
       blur: 'Glass blur intensity',
       blurDesc: 'Affects the inner backdrop-filter strength (the macOS system frosted layer can not be tuned at runtime).',
-      startupOpen: 'On launch',
-      startupOverview: 'Overview',
-      startupLast: 'Last position',
-      startupAtBoot: 'Launch at login',
     },
     about: {
       tagline: 'Speak naturally, write perfectly',
