+++ openless-all/app/src/i18n/ja.ts
@@ -497,17 +497,31 @@ export const ja: typeof zhCN = {
     kicker: 'SETTINGS',
     title: '設定',
     desc: '録音、プロバイダー、ショートカット、権限の設定。',
-    sections: {
-      recording: '録音',
-      providers: 'プロバイダー',
-      shortcuts: 'ショートカット',
-      permissions: '権限',
-      language: '言語',
-      advanced: '詳細設定',
-      about: '情報',
+    dataStorage: {
+      title: 'データ保存',
+      desc: 'この端末に保存される会話履歴とコンテキスト。',
+    },
+    debug: {
+      title: 'デバッグツール',
+      desc: '認識の問題を調査するときに使用。通常はオフのままで構いません。',
+    },
+    marketplace: {
+      title: '拡張マーケット',
+      desc: 'スタイルマーケットの投稿者 ID。スタイルの閲覧とインストールは「スタイル」ページで行います。',
+      github: {
+        signIn: 'GitHub でログイン',
+        signedIn: 'GitHub でログイン済み',
+        signedOut: 'ログインするとスタイルの投稿・いいねができます。',
+        signOut: 'ログアウト',
+        starting: 'ログインを開始しています…',
+        codeHint: '開いた GitHub ページでこのコードを入力してください：',
+        openGithub: 'GitHub を開く',
+        waiting: 'GitHub を開きました。承認するとログインします…',
+        failed: 'ログインに失敗しました。再試行してください',
+      },
     },
     recording: {
-      title: '録音',
+      title: '録音と入力',
       desc: 'グローバル録音のショートカットとトリガー方式を定義します。',
       hotkeyLabel: '録音ショートカット',
       hotkeyDescAcc: '押すと音声キャプチャを開始（グローバル）。アクセシビリティ権限が必要です。',
@@ -767,8 +781,9 @@ export const ja: typeof zhCN = {
       privacy: 'プライバシー',
       privacyDesc: 'すべてのデータはローカルに保存。クラウド API は録音を保持しません。',
       localFirst: 'ローカル優先',
+      linksTitle: 'ドキュメント',
       betaChannelLabel: 'Beta チャンネルに参加',
-      betaChannelDesc: '有効にすると Beta 更新を受信。不安定な場合があります。',
+      betaChannelDesc: '新機能をいち早く体験',
       betaChannelFetching: '最新 Beta 版を取得中…',
       betaChannelFetchBtn: '最新 Beta を確認',
       betaChannelLatestPrefix: '最新 Beta：',
@@ -809,25 +824,16 @@ export const ja: typeof zhCN = {
   },
   modal: {
     sections: {
-      account: 'アカウント',
-      settings: '設定',
+      general: '一般',
+      services: 'サービス',
+      privacy: 'プライバシー',
+      advanced: '詳細設定',
       personalize: 'パーソナライズ',
       about: '情報',
       helpCenter: 'ヘルプセンター',
       releaseNotes: 'リリースノート',
     },
-    account: {
-      localUser: 'ローカルユーザー',
-      localUserDesc: '未ログイン · すべてのデータはローカルに保存',
-      loginSync: 'ログイン / 同期',
-      footer: 'デフォルトで完全ローカル動作。ログインすると語彙とスタイルプリセットをデバイス間で同期。',
-    },
     personalize: {
-      appearance: '外観',
-      appearanceDesc: 'システムに従う / ライト / ダーク',
-      appearanceSystem: 'システムに従う',
-      appearanceLight: 'ライト',
-      appearanceDark: 'ダーク',
       font: 'フォントサイズ',
       fontDesc: 'UI のフォントサイズを全体的にスケール。即時反映。',
       fontSmall: '小',
@@ -835,10 +841,6 @@ export const ja: typeof zhCN = {
       fontLarge: '大',
       blur: 'すりガラス強度',
       blurDesc: 'ウィンドウ内側の backdrop-filter 強度に影響（macOS のシステムフロスト層が動かない場合に調整）。',
-      startupOpen: '起動時に開く',
-      startupOverview: '概要',
-      startupLast: '前回の位置',
-      startupAtBoot: '起動時に自動起動',
     },
     about: {
       tagline: '自然に話し、きれいに書く',
