+++ openless-all/app/src/i18n/ko.ts
@@ -497,17 +497,31 @@ export const ko: typeof zhCN = {
     kicker: 'SETTINGS',
     title: '설정',
     desc: '녹음, 공급자, 단축키, 권한 설정.',
-    sections: {
-      recording: '녹음',
-      providers: '공급자',
-      shortcuts: '단축키',
-      permissions: '권한',
-      language: '언어',
-      advanced: '고급',
-      about: '정보',
+    dataStorage: {
+      title: '데이터 저장',
+      desc: '이 기기에 보관되는 대화 기록과 컨텍스트.',
+    },
+    debug: {
+      title: '디버그 도구',
+      desc: '인식 문제를 진단할 때 사용합니다. 평소에는 꺼두어도 됩니다.',
+    },
+    marketplace: {
+      title: '확장 마켓',
+      desc: '스타일 마켓 업로드 신원. 스타일 둘러보기와 설치는 「스타일」 페이지에서 합니다.',
+      github: {
+        signIn: 'GitHub로 로그인',
+        signedIn: 'GitHub로 로그인됨',
+        signedOut: '로그인하면 스타일 업로드와 좋아요를 할 수 있습니다.',
+        signOut: '로그아웃',
+        starting: '로그인을 시작하는 중…',
+        codeHint: '열린 GitHub 페이지에서 이 코드를 입력하세요:',
+        openGithub: 'GitHub 열기',
+        waiting: 'GitHub를 열었습니다. 승인하면 로그인됩니다…',
+        failed: '로그인 실패, 다시 시도하세요',
+      },
     },
     recording: {
-      title: '녹음',
+      title: '녹음 및 입력',
       desc: '전역 녹음의 단축키와 트리거 방식을 정의합니다.',
       hotkeyLabel: '녹음 단축키',
       hotkeyDescAcc: '누르면 음성 캡처 시작(전역). 접근성 권한이 필요합니다.',
@@ -767,8 +781,9 @@ export const ko: typeof zhCN = {
       privacy: '프라이버시',
       privacyDesc: '모든 데이터는 로컬에 저장됩니다. 클라우드 API 는 녹음을 보관하지 않습니다.',
       localFirst: '로컬 우선',
+      linksTitle: '문서 링크',
       betaChannelLabel: 'Beta 채널 참여',
-      betaChannelDesc: '활성화 시 Beta 업데이트 수신. 불안정할 수 있으며 얼리 어답터만 권장.',
+      betaChannelDesc: '새 기능 미리 체험',
       betaChannelFetching: '최신 Beta 버전을 가져오는 중…',
       betaChannelFetchBtn: '최신 Beta 확인',
       betaChannelLatestPrefix: '최신 Beta:',
@@ -809,25 +824,16 @@ export const ko: typeof zhCN = {
   },
   modal: {
     sections: {
-      account: '계정',
-      settings: '설정',
+      general: '일반',
+      services: '서비스',
+      privacy: '프라이버시',
+      advanced: '고급',
       personalize: '개인 설정',
       about: '정보',
       helpCenter: '도움말 센터',
       releaseNotes: '릴리스 노트',
     },
-    account: {
-      localUser: '로컬 사용자',
-      localUserDesc: '로그인하지 않음 · 모든 데이터는 로컬에 저장',
-      loginSync: '로그인 / 동기화',
-      footer: '기본적으로 완전 로컬 실행. 로그인하면 어휘와 스타일 프리셋을 기기 간 동기화.',
-    },
     personalize: {
-      appearance: '모양',
-      appearanceDesc: '시스템 따라가기 / 라이트 / 다크',
-      appearanceSystem: '시스템 따라가기',
-      appearanceLight: '라이트',
-      appearanceDark: '다크',
       font: '글꼴 크기',
       fontDesc: 'UI 글꼴 크기를 전체 스케일. 즉시 반영.',
       fontSmall: '소',
@@ -835,10 +841,6 @@ export const ko: typeof zhCN = {
       fontLarge: '대',
       blur: '서리유리 강도',
       blurDesc: '창 내부 backdrop-filter 강도에 영향(macOS 시스템 서리 레이어가 작동하지 않을 때 조정).',
-      startupOpen: '시작 시 열기',
-      startupOverview: '개요',
-      startupLast: '마지막 위치',
-      startupAtBoot: '부팅 시 자동 시작',
     },
     about: {
       tagline: '자연스럽게 말하고, 정확하게 작성하세요',
