+++ openless-all/app/src/i18n/zh-CN.ts
@@ -493,17 +493,31 @@ export const zhCN = {
     kicker: 'SETTINGS',
     title: '设置',
     desc: '录音、提供商、快捷键与权限配置。',
-    sections: {
-      recording: '录音',
-      providers: '提供商',
-      shortcuts: '快捷键',
-      permissions: '权限',
-      language: '语言',
-      advanced: '高级',
-      about: '关于',
+    dataStorage: {
+      title: '数据存储',
+      desc: '本机保留的历史会话与对话上下文。',
+    },
+    debug: {
+      title: '调试工具',
+      desc: '排查识别问题时使用，平时无需开启。',
+    },
+    marketplace: {
+      title: '扩展市场',
+      desc: '风格市场的上传身份。浏览与安装风格在「风格」页内完成。',
+      github: {
+        signIn: '用 GitHub 账号登录',
+        signedIn: '已通过 GitHub 登录',
+        signedOut: '登录后即可上传风格、给风格点赞。',
+        signOut: '退出登录',
+        starting: '正在发起登录…',
+        codeHint: '在打开的 GitHub 页面输入这个验证码：',
+        openGithub: '打开 GitHub',
+        waiting: '已打开 GitHub，完成授权后会自动登录…',
+        failed: '登录失败，请重试',
+      },
     },
     recording: {
-      title: '录音',
+      title: '录音与输入',
       desc: '全局录音的快捷键与触发方式。',
       hotkeyLabel: '录音快捷键',
       hotkeyDescAcc: '按下开始捕获语音，全局生效（需辅助功能权限）。',
@@ -763,8 +777,9 @@ export const zhCN = {
       privacy: '隐私',
       privacyDesc: '所有数据仅保存在本机，云端 API 不保留录音。',
       localFirst: '本地优先',
+      linksTitle: '文档链接',
       betaChannelLabel: '加入 Beta 渠道',
-      betaChannelDesc: '开启后自动接收 Beta 版本。可能不稳定，仅推荐愿意尝鲜的用户。',
+      betaChannelDesc: '抢先体验新功能',
       betaChannelFetching: '正在获取最新 Beta 版本…',
       betaChannelFetchBtn: '查询最新 Beta',
       betaChannelLatestPrefix: '最新 Beta：',
@@ -805,25 +820,16 @@ export const zhCN = {
   },
   modal: {
     sections: {
-      account: '账户',
-      settings: '设置',
+      general: '通用',
+      services: '服务',
+      privacy: '隐私',
+      advanced: '高级',
       personalize: '个性化',
       about: '关于',
       helpCenter: '帮助中心',
-      releaseNotes: '版本说明',
-    },
-    account: {
-      localUser: '本地用户',
-      localUserDesc: '未登录 · 所有数据保存在本机',
-      loginSync: '登录 / 同步',
-      footer: '默认完全本地运行。登录后可跨设备同步词汇表与风格预设。',
+      releaseNotes: '发布日志',
     },
     personalize: {
-      appearance: '外观',
-      appearanceDesc: '跟随系统 / 浅色 / 深色',
-      appearanceSystem: '跟随系统',
-      appearanceLight: '浅色',
-      appearanceDark: '深色',
       font: '字体大小',
       fontDesc: '整体缩放界面字号，立即生效。',
       fontSmall: '小',
@@ -831,10 +837,6 @@ export const zhCN = {
       fontLarge: '大',
       blur: '毛玻璃强度',
       blurDesc: '影响窗口内层 backdrop-filter 强度（macOS 系统磨砂层无法运行时调）。',
-      startupOpen: '启动时打开',
-      startupOverview: '概览',
-      startupLast: '上次位置',
-      startupAtBoot: '开机自启',
     },
     about: {
       tagline: '自然说话，完美书写',
