+++ openless-all/app/src/i18n/zh-TW.ts
@@ -495,17 +495,31 @@ export const zhTW: typeof zhCN = {
     kicker: 'SETTINGS',
     title: '設置',
     desc: '錄音、提供商、快捷鍵與權限配置。',
-    sections: {
-      recording: '錄音',
-      providers: '提供商',
-      shortcuts: '快捷鍵',
-      permissions: '權限',
-      language: '語言',
-      advanced: '高級',
-      about: '關於',
+    dataStorage: {
+      title: '資料儲存',
+      desc: '本機保留的歷史會話與對話上下文。',
+    },
+    debug: {
+      title: '除錯工具',
+      desc: '排查辨識問題時使用，平時無需開啟。',
+    },
+    marketplace: {
+      title: '擴充市集',
+      desc: '風格市集的上傳身份。瀏覽與安裝風格在「風格」頁內完成。',
+      github: {
+        signIn: '用 GitHub 帳號登入',
+        signedIn: '已透過 GitHub 登入',
+        signedOut: '登入後即可上傳風格、為風格按讚。',
+        signOut: '登出',
+        starting: '正在發起登入…',
+        codeHint: '在開啟的 GitHub 頁面輸入這個驗證碼：',
+        openGithub: '開啟 GitHub',
+        waiting: '已開啟 GitHub，完成授權後會自動登入…',
+        failed: '登入失敗，請重試',
+      },
     },
     recording: {
-      title: '錄音',
+      title: '錄音與輸入',
       desc: '定義全局錄音的快捷鍵與觸發方式。',
       hotkeyLabel: '錄音快捷鍵',
       hotkeyDescAcc: '按下即開始捕獲語音，全局生效。需要授予輔助功能權限。',
@@ -765,8 +779,9 @@ export const zhTW: typeof zhCN = {
       privacy: '隱私',
       privacyDesc: '所有資料僅保存在本機，雲端 API 不保留錄音。',
       localFirst: '本地優先',
+      linksTitle: '文件連結',
       betaChannelLabel: '加入 Beta 渠道',
-      betaChannelDesc: '開啟後自動接收 Beta 版本。可能不穩定，僅推薦願意嘗鮮的用戶。',
+      betaChannelDesc: '搶先體驗新功能',
       betaChannelFetching: '正在獲取最新 Beta 版本…',
       betaChannelFetchBtn: '查詢最新 Beta',
       betaChannelLatestPrefix: '最新 Beta：',
@@ -807,25 +822,16 @@ export const zhTW: typeof zhCN = {
   },
   modal: {
     sections: {
-      account: '賬戶',
-      settings: '設置',
+      general: '通用',
+      services: '服務',
+      privacy: '隱私',
+      advanced: '高級',
       personalize: '個性化',
       about: '關於',
       helpCenter: '幫助中心',
-      releaseNotes: '版本說明',
-    },
-    account: {
-      localUser: '本地用戶',
-      localUserDesc: '未登錄 · 所有數據保存在本機',
-      loginSync: '登錄 / 同步',
-      footer: '預設完全本地運行。登入後可跨裝置同步詞彙表與風格預設。',
+      releaseNotes: '發佈日誌',
     },
     personalize: {
-      appearance: '外觀',
-      appearanceDesc: '跟隨系統 / 淺色 / 深色',
-      appearanceSystem: '跟隨系統',
-      appearanceLight: '淺色',
-      appearanceDark: '深色',
       font: '字體大小',
       fontDesc: '整體縮放界面字號，立即生效。',
       fontSmall: '小',
@@ -833,10 +839,6 @@ export const zhTW: typeof zhCN = {
       fontLarge: '大',
       blur: '毛玻璃強度',
       blurDesc: '影響窗口內層 backdrop-filter 強度（macOS 系統磨砂層無法運行時調）。',
-      startupOpen: '啓動時打開',
-      startupOverview: '概覽',
-      startupLast: '上次位置',
-      startupAtBoot: '開機自啓',
     },
     about: {
       tagline: '自然說話，完美書寫',
