+++ openless-all/app/src/pages/settings/ProvidersSection.tsx
@@ -1,0 +1,772 @@
+// 服务 → AI 提供商：LLM 润色模型 + ASR 语音转写两张卡片。
+// 自 Settings.tsx 整体迁出，逻辑零改动；i18n key 全部保持 `settings.providers.*`。
+
+import { useEffect, useRef, useState, type CSSProperties, type ReactNode } from 'react';
+import { useTranslation } from 'react-i18next';
+import { Icon } from '../../components/Icon';
+import { detectOS } from '../../components/WindowChrome';
+import {
+  listProviderModels,
+  readCredential,
+  setActiveAsrProvider,
+  setActiveLlmProvider,
+  setCredential,
+  validateProviderCredentials,
+} from '../../lib/ipc';
+import { emitSaved } from '../../lib/savedEvent';
+import { useHotkeySettings } from '../../state/HotkeySettingsContext';
+import { SelectLite } from '../../components/ui/SelectLite';
+import { Card } from '../_atoms';
+import { SettingRow, SectionTitle, Toggle, inputStyle, type AsrPresetId } from './shared';
+
+function LlmThinkingToggle({ enabled, onToggle }: { enabled: boolean; onToggle: (next: boolean) => void }) {
+  const { t } = useTranslation();
+  return (
+    <div
+      title={t('settings.providers.thinkingModeHint')}
+      style={{
+        display: 'flex',
+        alignItems: 'center',
+        gap: 6,
+        paddingLeft: 2,
+        whiteSpace: 'nowrap',
+      }}
+    >
+      <span style={{ fontSize: 11.5, color: 'var(--ol-ink-4)' }}>
+        {t('settings.providers.thinkingModeLabel')}
+      </span>
+      <Toggle on={enabled} onToggle={onToggle} />
+      <span style={{ fontSize: 11.5, color: enabled ? 'var(--ol-blue)' : 'var(--ol-ink-4)' }}>
+        {enabled ? t('settings.providers.thinkingModeOn') : t('settings.providers.thinkingModeOff')}
+      </span>
+    </div>
+  );
+}
+
+const LLM_PRESETS = [
+  {
+    id: 'ark',
+    nameKey: 'ark',
+    baseUrl: 'https://ark.cn-beijing.volces.com/api/v3',
+    modelPlaceholder: 'deepseek-v3-2',
+  },
+  {
+    id: 'deepseek',
+    nameKey: 'deepseek',
+    baseUrl: 'https://api.deepseek.com/v1',
+    modelPlaceholder: 'deepseek-v4-flash',
+  },
+  {
+    id: 'siliconflow',
+    nameKey: 'siliconflow',
+    baseUrl: 'https://api.siliconflow.cn/v1',
+    modelPlaceholder: 'Qwen/Qwen2.5-7B-Instruct',
+  },
+  {
+    id: 'openai',
+    nameKey: 'openai',
+    baseUrl: 'https://api.openai.com/v1',
+    modelPlaceholder: 'gpt-4o',
+  },
+  {
+    // 谷歌官方 Gemini API（原生 generateContent，不走 OpenAI 兼容 shim）。
+    // baseUrl 末尾 /v1beta 是当前 Generally Available 的 path（ai.google.dev/api）。
+    // 后端 llm_gemini.rs 会拼成 `{baseUrl}/models/{model}:generateContent`，
+    // 并按 Gemini 原生通道级 thinkingConfig 关闭或压低思考，不在前端维护模型适配表。
+    // 模型列表用 ProviderTools「拉取模型」按钮取，
+    // 由 commands.rs::fetch_provider_models 识别 generativelanguage 域名后按 Gemini shape 解析。
+    id: 'gemini',
+    nameKey: 'gemini',
+    baseUrl: 'https://generativelanguage.googleapis.com/v1beta',
+    modelPlaceholder: 'gemini-2.5-flash',
+  },
+  {
+    id: 'codex_oauth',
+    nameKey: 'codexOAuth',
+    baseUrl: '',
+    modelPlaceholder: 'gpt-5.3-codex-spark',
+  },
+  {
+    id: 'mimo',
+    nameKey: 'mimo',
+    baseUrl: 'https://api.xiaomimimo.com/v1',
+    modelPlaceholder: 'xiaomi/mimo-v2-flash',
+  },
+  {
+    id: 'cometapi',
+    nameKey: 'cometapi',
+    baseUrl: 'https://api.cometapi.com/v1',
+    modelPlaceholder: 'gpt-4o',
+  },
+  {
+    id: 'openrouterFree',
+    nameKey: 'openrouterFree',
+    baseUrl: 'https://openrouter.ai/api/v1',
+    modelPlaceholder: 'qwen/qwen3-coder:free',
+  },
+  {
+    id: 'alibabaCoding',
+    nameKey: 'alibabaCoding',
+    baseUrl: 'https://coding-intl.dashscope.aliyuncs.com/v1',
+    modelPlaceholder: 'qwen3-coder-plus',
+  },
+  {
+    id: 'codingPlanX',
+    nameKey: 'codingPlanX',
+    baseUrl: 'https://api.codingplanx.ai/v1',
+    modelPlaceholder: 'gpt-5-mini',
+  },
+  {
+    id: 'custom',
+    nameKey: 'custom',
+    baseUrl: '',
+    modelPlaceholder: '',
+  },
+] as const;
+
+type LlmPresetId = typeof LLM_PRESETS[number]['id'];
+
+const ASR_DEFAULT_RESOURCE_ID = 'volc.bigasr.sauc.duration';
+
+// `volcengine` / `bailian` 走自建流式客户端；其余走 OpenAI 兼容
+// `/audio/transcriptions`（`coordinator.rs::is_whisper_compatible_provider`）。
+// 新增兼容厂商：
+//   1. 在这里加一项 `{ id, nameKey, baseUrl, model }`；
+//   2. `coordinator.rs::is_whisper_compatible_provider` 加同名 id；
+//   3. 在 i18n 的 `settings.providers.presets.<nameKey>` 加文案。
+// `AsrPresetId` 定义在 settings/shared.tsx，LocalModelSection / ProvidersSection 共用同一份。
+const ASR_PRESETS: ReadonlyArray<{ id: AsrPresetId; nameKey: string; baseUrl: string; model: string }> = [
+  { id: 'volcengine',   nameKey: 'asrVolcengine',   baseUrl: '',                                              model: ''                              },
+  { id: 'bailian',      nameKey: 'asrBailian',     baseUrl: 'wss://dashscope.aliyuncs.com/api-ws/v1/inference/', model: 'fun-asr-realtime'             },
+  { id: 'siliconflow',  nameKey: 'asrSiliconflow',  baseUrl: 'https://api.siliconflow.cn/v1',                  model: 'FunAudioLLM/SenseVoiceSmall' },
+  { id: 'zhipu',        nameKey: 'asrZhipu',        baseUrl: 'https://open.bigmodel.cn/api/paas/v4',           model: 'glm-asr-2512'                },
+  { id: 'groq',         nameKey: 'asrGroq',         baseUrl: 'https://api.groq.com/openai/v1',                 model: 'whisper-large-v3-turbo'      },
+  { id: 'whisper',      nameKey: 'asrWhisper',      baseUrl: 'https://api.openai.com/v1',                      model: 'whisper-1'                   },
+  { id: 'foundry-local-whisper', nameKey: 'asrFoundryLocalWhisper', baseUrl: '',                              model: ''                              },
+  // 本地 Qwen3-ASR：无 baseUrl/model 配置，模型在「高级 → 本地模型」里下载与切换。
+  { id: 'local-qwen3',  nameKey: 'asrLocalQwen3',   baseUrl: '',                                              model: ''                              },
+];
+
+export function ProvidersSection() {
+  const { t } = useTranslation();
+  const { prefs, updatePrefs } = useHotkeySettings();
+  // `*Provider` 立即跟随 <select> 改动（受控组件必须实时反映用户输入）；
+  // `committed*Provider` 才决定 CredentialField 的 key，仅在后端 active
+  // 切换 + 默认值写完后再 commit。两者拆开是为了同时满足：
+  //   - <select> 立刻显示用户的选择（issue #220 P2：codex 指出受控选不应等 await）
+  //   - CredentialField 不要在后端 active 切完前 remount（issue #219：避免读到旧 entry）
+  // `*SwitchSeq` 是 stale-write 守卫：用户 100ms 内连点两次时，先发的请求晚到不
+  // 会覆盖后发的 commit。
+  const [llmProvider, setLlmProvider] = useState<LlmPresetId>('ark');
+  const [asrProvider, setAsrProvider] = useState<AsrPresetId>('volcengine');
+  const [committedLlmProvider, setCommittedLlmProvider] = useState<LlmPresetId>('ark');
+  const [committedAsrProvider, setCommittedAsrProvider] = useState<AsrPresetId>('volcengine');
+  const llmSwitchSeqRef = useRef(0);
+  const asrSwitchSeqRef = useRef(0);
+  const [llmModelRevision, setLlmModelRevision] = useState(0);
+  const [asrModelRevision, setAsrModelRevision] = useState(0);
+  const os = detectOS();
+  // 主 ASR 下拉只列云端选项；本地推理（local-qwen3 / foundry-local-whisper）
+  // 移到「高级 → 本地模型」，防止新手误开 CPU 推理。
+  const visibleAsrPresets = ASR_PRESETS.filter(
+    p => p.id !== 'foundry-local-whisper' && p.id !== 'local-qwen3',
+  );
+
+  useEffect(() => {
+    if (!prefs) return;
+    const knownLlm = LLM_PRESETS.find(x => x.id === prefs.activeLlmProvider);
+    const llmId = knownLlm ? knownLlm.id : 'custom';
+    setLlmProvider(llmId);
+    setCommittedLlmProvider(llmId);
+    // ASR 在 ALL ASR_PRESETS 里查（不是 visibleAsrPresets）——本地选项虽然
+    // 从下拉里藏起来了，但若用户曾在「高级」里启用过 local-qwen3，主 Card
+    // 仍要识别出 active 是本地，并切到「正在使用本地 ASR」的 notice 渲染。
+    const knownAsr = ASR_PRESETS.find(x => x.id === prefs.activeAsrProvider);
+    const asrId = knownAsr ? knownAsr.id : 'volcengine';
+    setAsrProvider(asrId);
+    setCommittedAsrProvider(asrId);
+  }, [prefs, os]);
+
+  // issue #219 / #220 P2：
+  //   1. 立刻 setLlmProvider —— 受控 <select> 必须反映用户最新选择。
+  //   2. 用 seq 守卫每个 await：用户连点两次时旧请求晚到也不会盖掉新选择。
+  //   3. 仅 setCommittedLlmProvider 之后 CredentialField 才 remount 读新 entry，
+  //      此时后端 root.active.llm 已经是 id，lookup_account 落到正确 entry。
+  //   4. endpoint/model 默认值仅在该 provider entry 该字段为空时才填，不覆盖用户自定义。
+  const onLlmProviderChange = async (id: LlmPresetId) => {
+    setLlmProvider(id);
+    const seq = ++llmSwitchSeqRef.current;
+    emitSaved('saving', t('common.saving'));
+    try {
+      await setActiveLlmProvider(id);
+      if (seq !== llmSwitchSeqRef.current) return;
+      if (prefs) {
+        const next = { ...prefs, activeLlmProvider: id };
+        await updatePrefs(next);
+        if (seq !== llmSwitchSeqRef.current) return;
+      }
+      const preset = LLM_PRESETS.find(p => p.id === id);
+      // 修 bug：所有 LLM provider 共用 `ark.endpoint` / `ark.model_id` 一对凭据槽
+      // （persistence.rs 没做 per-provider 隔离）。旧逻辑只在槽空时填默认值，
+      // 老用户切换 preset 时槽里早有旧值——dropdown 看着切了，polish 实际还是
+      // 打老 endpoint。改成：切到任何非 custom 预设都强制覆盖 endpoint 与 model
+      // 到该预设的默认值，让"切换"真切到位。custom 预设没有默认值，跳过。
+      if (preset && preset.id !== 'custom') {
+        if (preset.baseUrl) {
+          await setCredential('ark.endpoint', preset.baseUrl);
+          if (seq !== llmSwitchSeqRef.current) return;
+        }
+        if (preset.modelPlaceholder) {
+          await setCredential('ark.model_id', preset.modelPlaceholder);
+          if (seq !== llmSwitchSeqRef.current) return;
+        }
+      }
+      setCommittedLlmProvider(id);
+      emitSaved('saved', t('common.saved'));
+    } catch (err) {
+      // seq 守卫：只有当前 call 还是最新时才把 saving 翻成 failed；
+      // 旧 call 早被 newer call 的 emitSaved('saving') 覆盖，再叠 failed 会
+      // 把 newer 正在跑的 saving 假伪成失败。
+      if (seq === llmSwitchSeqRef.current) {
+        emitSaved('failed', t('common.operationFailed'));
+      }
+      throw err;
+    }
+  };
+
+  const onLlmThinkingToggle = (enabled: boolean) => {
+    if (!prefs) return;
+    void updatePrefs(current => ({ ...current, llmThinkingEnabled: enabled })).catch(error => {
+      console.error('[settings] failed to update LLM thinking mode', error);
+      emitSaved('failed', t('common.operationFailed'));
+    });
+  };
+
+  const onAsrProviderChange = async (id: AsrPresetId) => {
+    setAsrProvider(id);
+    const seq = ++asrSwitchSeqRef.current;
+    emitSaved('saving', t('common.saving'));
+    try {
+      await setActiveAsrProvider(id);
+      if (seq !== asrSwitchSeqRef.current) return;
+      if (prefs) {
+        const next = { ...prefs, activeAsrProvider: id };
+        await updatePrefs(next);
+        if (seq !== asrSwitchSeqRef.current) return;
+      }
+      // OpenAI 兼容厂商首次切换时预填 baseUrl / model 默认值，省得用户必踩
+      // 「跨厂商 model 名根本不一样」的坑；但用户已自定义后就不再覆盖。
+      // volcengine 走另一套凭据，跳过。
+      const preset = ASR_PRESETS.find(p => p.id === id);
+      if (preset && preset.baseUrl) {
+        const existing = await readCredential('asr.endpoint');
+        if (seq !== asrSwitchSeqRef.current) return;
+        if (!existing) {
+          await setCredential('asr.endpoint', preset.baseUrl);
+          if (seq !== asrSwitchSeqRef.current) return;
+        }
+      }
+      if (preset && preset.model) {
+        const existing = await readCredential('asr.model');
+        if (seq !== asrSwitchSeqRef.current) return;
+        if (!existing) {
+          await setCredential('asr.model', preset.model);
+          if (seq !== asrSwitchSeqRef.current) return;
+        }
+      }
+      setCommittedAsrProvider(id);
+      emitSaved('saved', t('common.saved'));
+    } catch (err) {
+      // seq 守卫同上 onLlmProviderChange：旧 call 不要把 newer call 的 saving
+      // 伪造成 failed。
+      if (seq === asrSwitchSeqRef.current) {
+        emitSaved('failed', t('common.operationFailed'));
+      }
+      throw err;
+    }
+  };
+
+  // preset 决定 placeholder 与 default —— 必须跟着 committed*Provider 走，
+  // 否则受控 <select> 立刻切到新厂商，但凭据字段还在显示旧 entry，placeholder
+  // 会先于实际数据切换、视觉上对不上。
+  const preset = LLM_PRESETS.find(p => p.id === committedLlmProvider) ?? LLM_PRESETS[LLM_PRESETS.length - 1];
+  const codexOAuthSelected = committedLlmProvider === 'codex_oauth';
+  const asrPreset = visibleAsrPresets.find(p => p.id === committedAsrProvider);
+  return (
+    <>
+      <div style={{ fontSize: 11.5, color: 'var(--ol-ink-4)', lineHeight: 1.6, marginBottom: 10 }}>
+        {t('settings.providers.credentialStorageNotice')}
+      </div>
+      <Card>
+        <div style={{ marginBottom: 10 }}>
+          <SectionTitle>{t('settings.providers.llmTitle')}</SectionTitle>
+        </div>
+        {/* desc 已去掉——'选择后将自动填入 Base URL 默认值' 在 180px label 列必换行成两行，
+            视觉上 label 区出现"字体单独占一行"。下拉自身已经表达了"切换"含义，desc 冗余。 */}
+        <SettingRow label={t('settings.providers.providerLabel')}>
+          <SelectLite
+            value={llmProvider}
+            onChange={next => onLlmProviderChange(next as LlmPresetId)}
+            options={LLM_PRESETS.map(p => ({
+              value: p.id,
+              label: t(`settings.providers.presets.${p.nameKey}`),
+            }))}
+            ariaLabel={t('settings.providers.providerLabel')}
+            style={{ ...inputStyle, width: '100%', maxWidth: 200 }}
+          />
+        </SettingRow>
+        {codexOAuthSelected ? (
+          <div style={{ fontSize: 11.5, color: 'var(--ol-ink-4)', lineHeight: 1.6, margin: '2px 0 10px' }}>
+            {t('settings.providers.codexOAuthNotice')}
+          </div>
+        ) : (
+          <>
+            <CredentialField key={`${committedLlmProvider}:api_key`} label={t('settings.providers.apiKeyLabel')} account="ark.api_key" mono mask />
+            <CredentialField key={`${committedLlmProvider}:endpoint`} label={t('settings.providers.baseUrlLabel')} account="ark.endpoint"
+              placeholder={preset.baseUrl || 'https://your-endpoint/v1'} />
+          </>
+        )}
+        <CredentialField key={`${committedLlmProvider}:model:${llmModelRevision}`} label={t('settings.providers.modelLabel')} account="ark.model_id"
+          placeholder={preset.modelPlaceholder || 'model-name'} mono
+          trailing={(
+            <LlmThinkingToggle
+              enabled={prefs?.llmThinkingEnabled ?? false}
+              onToggle={onLlmThinkingToggle}
+            />
+          )}
+        />
+        <ProviderTools key={committedLlmProvider} kind="llm" modelAccount="ark.model_id" onModelSelected={() => setLlmModelRevision(v => v + 1)} />
+      </Card>
+
+      <Card>
+        <div style={{ marginBottom: 10 }}>
+          <SectionTitle>{t('settings.providers.asrTitle')}</SectionTitle>
+        </div>
+        {/* 下拉只放云端选项；本地引擎激活时锁住 + 在下方放一行"ASR 提供商已被接管"提示，
+            未激活时不显示提示。 */}
+        <SettingRow label={t('settings.providers.providerLabel')}>
+          {(() => {
+            const isLocked =
+              committedAsrProvider === 'local-qwen3' ||
+              committedAsrProvider === 'foundry-local-whisper';
+            const selectedValue: AsrPresetId = isLocked ? committedAsrProvider : asrProvider;
+            // 跨机器同步异常兜底：committed 是本地但不在 visibleAsrPresets 里时，受控
+            // select 会回退到首项造成假象 —— 补一个 disabled option 让 select 找到当前值。
+            const anomalousLocal: AsrPresetId | null =
+              isLocked && !visibleAsrPresets.some(p => p.id === committedAsrProvider)
+                ? committedAsrProvider
+                : null;
+            const anomalousNameKey = anomalousLocal === 'local-qwen3'
+              ? 'asrLocalQwen3'
+              : anomalousLocal === 'foundry-local-whisper'
+                ? 'asrFoundryLocalWhisper'
+                : null;
+            return (
+              <div style={{ display: 'flex', flexDirection: 'column', gap: 6, alignItems: 'flex-start', minWidth: 0 }}>
+                <SelectLite
+                  value={selectedValue}
+                  disabled={isLocked}
+                  onChange={next => onAsrProviderChange(next as AsrPresetId)}
+                  options={[
+                    ...visibleAsrPresets.map(p => ({
+                      value: p.id,
+                      label: t(`settings.providers.presets.${p.nameKey}`),
+                    })),
+                    ...(anomalousLocal && anomalousNameKey
+                      ? [{
+                          value: anomalousLocal,
+                          label: t(`settings.providers.presets.${anomalousNameKey}`),
+                          disabled: true,
+                        }]
+                      : []),
+                  ]}
+                  ariaLabel={t('settings.providers.providerLabel')}
+                  style={{ ...inputStyle, width: '100%', maxWidth: 200 }}
+                />
+                {isLocked && (
+                  <div style={{ fontSize: 11, color: 'var(--ol-ink-4)', lineHeight: 1.5 }}>
+                    {t('settings.providers.asrProviderTakenOver')}
+                  </div>
+                )}
+              </div>
+            );
+          })()}
+        </SettingRow>
+        {committedAsrProvider === 'volcengine' ? (
+          <>
+            <CredentialField
+              key={`${committedAsrProvider}:app_key`}
+              label={t('settings.providers.volcengineAppKeyLabel')}
+              account="volcengine.app_key"
+              mono
+              mask
+            />
+            <CredentialField
+              key={`${committedAsrProvider}:access_key`}
+              label={t('settings.providers.volcengineAccessKeyLabel')}
+              account="volcengine.access_key"
+              mono
+              mask
+            />
+            <CredentialField
+              key={`${committedAsrProvider}:resource_id`}
+              label={t('settings.providers.volcengineResourceIdLabel')}
+              account="volcengine.resource_id"
+              mono
+              placeholder={ASR_DEFAULT_RESOURCE_ID} defaultValue={ASR_DEFAULT_RESOURCE_ID} />
+            <div style={{ marginTop: 2, fontSize: 11.5, color: 'var(--ol-ink-4)', lineHeight: 1.6 }}>
+              {t('settings.providers.volcengineMappingNote')}
+            </div>
+          </>
+        ) : committedAsrProvider === 'local-qwen3' || committedAsrProvider === 'foundry-local-whisper' ? (
+          // 用户已经在用本地 ASR——dropdown 行的 asrProviderTakenOver 已经把
+          // "在高级中切换或禁用"讲清楚了，body 不再重复。
+          // 模型管理 UI 唯一入口在「高级 → 本地模型」里的 <LocalAsr embedded />。
+          null
+        ) : (
+          <>
+            <CredentialField key={`${committedAsrProvider}:api_key`} label={t('settings.providers.apiKeyLabel')} account="asr.api_key" mono mask />
+            <CredentialField key={`${committedAsrProvider}:endpoint`} label={t('settings.providers.baseUrlLabel')} account="asr.endpoint"
+              placeholder={asrPreset?.baseUrl || 'https://api.openai.com/v1'}
+              defaultValue={asrPreset?.baseUrl || undefined} />
+            <CredentialField key={`${committedAsrProvider}:model:${asrModelRevision}`} label={t('settings.providers.modelLabel')} account="asr.model"
+              placeholder={asrPreset?.model || 'whisper-1'} />
+            {committedAsrProvider === 'bailian' && (
+              <>
+                <CredentialField
+                  key={`${committedAsrProvider}:vocabulary_id`}
+                  label={t('settings.providers.bailianVocabularyIdLabel')}
+                  account="asr.vocabulary_id"
+                  mono
+                  placeholder="vocab-..."
+                />
+                <div style={{ marginTop: 2, fontSize: 11.5, color: 'var(--ol-ink-4)', lineHeight: 1.6 }}>
+                  {t('settings.providers.bailianVocabularyIdNote')}
+                </div>
+              </>
+            )}
+            <ProviderTools kind="asr" modelAccount="asr.model" onModelSelected={() => setAsrModelRevision(v => v + 1)} />
+          </>
+        )}
+      </Card>
+    </>
+  );
+}
+
+type ProviderToolStatus = 'idle' | 'loading' | 'success' | 'empty' | 'error';
+
+function ProviderTools({ kind, modelAccount, onModelSelected }: { kind: 'llm' | 'asr'; modelAccount: string; onModelSelected: () => void }) {
+  const { t } = useTranslation();
+  const [models, setModels] = useState<string[]>([]);
+  const [selectedModel, setSelectedModel] = useState('');
+  const [status, setStatus] = useState<ProviderToolStatus>('idle');
+  const [message, setMessage] = useState('');
+
+  const setResult = (next: ProviderToolStatus, nextMessage: string) => {
+    setStatus(next);
+    setMessage(nextMessage);
+  };
+
+  const validate = async () => {
+    setModels([]);
+    setSelectedModel('');
+    setResult('loading', t('settings.providers.validating'));
+    try {
+      const result = await validateProviderCredentials(kind);
+      setResult(result.ok ? 'success' : 'error', t('settings.providers.validateSuccess'));
+    } catch (error) {
+      const message = error instanceof Error ? error.message : String(error);
+      if ((kind === 'llm' && message === 'llmModelMissing') || (kind === 'asr' && message === 'asrModelMissing')) {
+        setResult('empty', t('settings.providers.modelMissing'));
+        return;
+      }
+      if (message === 'modelsEmpty') {
+        setResult('empty', t('settings.providers.modelsEmpty'));
+        return;
+      }
+      setResult('error', providerErrorMessage(error, t));
+    }
+  };
+
+  const loadModels = async () => {
+    setResult('loading', t('settings.providers.loadingModels'));
+    try {
+      const result = await listProviderModels(kind);
+      setModels(result.models);
+      if (result.models.length === 0) {
+        setResult('empty', t('settings.providers.modelsEmpty'));
+      } else {
+        setSelectedModel('');
+        setResult('success', t('settings.providers.modelsLoaded', { count: result.models.length }));
+      }
+    } catch (error) {
+      setModels([]);
+      setResult('error', providerErrorMessage(error, t));
+    }
+  };
+
+  const applyModel = async (model: string) => {
+    setResult('loading', t('common.saving'));
+    try {
+      await setCredential(modelAccount, model);
+      setSelectedModel(model);
+      onModelSelected();
+      setResult('success', t('settings.providers.modelSaved', { model }));
+    } catch (error) {
+      setResult('error', providerErrorMessage(error, t));
+    }
+  };
+
+  return (
+    <SettingRow label={t('settings.providers.toolsLabel')}>
+      <div style={{ display: 'flex', flexDirection: 'column', gap: 8, width: '100%', maxWidth: 420 }}>
+        <div style={{ display: 'flex', gap: 6, alignItems: 'center', flexWrap: 'wrap' }}>
+          <button onClick={validate} style={miniBtnStyle} disabled={status === 'loading'}>{t('settings.providers.validate')}</button>
+          <button onClick={loadModels} style={miniBtnStyle} disabled={status === 'loading'}>{t('settings.providers.fetchModels')}</button>
+          {models.length > 0 && (
+            <SelectLite
+              value={selectedModel}
+              onChange={applyModel}
+              disabled={status === 'loading'}
+              options={models.map(model => ({ value: model, label: model }))}
+              placeholder={t('settings.providers.selectModel')}
+              ariaLabel={t('settings.providers.selectModel')}
+              style={{ ...inputStyle, maxWidth: 220 }}
+            />
+          )}
+        </div>
+        {message && (
+          <span style={{ fontSize: 11, color: status === 'error' ? 'var(--ol-warn)' : status === 'empty' ? 'var(--ol-ink-4)' : 'var(--ol-ok)', lineHeight: 1.4 }}>
+            {message}
+          </span>
+        )}
+      </div>
+    </SettingRow>
+  );
+}
+
+function providerErrorMessage(error: unknown, t: ReturnType<typeof useTranslation>['t']): string {
+  const message = error instanceof Error ? error.message : String(error);
+  if (message.startsWith('providerHttpStatus:')) {
+    return t('settings.providers.providerHttpStatus', { status: message.split(':')[1] || '?' });
+  }
+  if (message === 'endpointMustUseHttps') return t('settings.providers.endpointMustUseHttps');
+  if (message === 'endpointInvalid') return t('settings.providers.endpointInvalid');
+  if (message === 'providerResponseTooLarge') return t('settings.providers.responseTooLarge');
+  if (message === 'asrInvalidJson') return t('settings.providers.asrInvalidJson');
+  if (message === 'asrMissingTextField') return t('settings.providers.asrMissingTextField');
+  if (message === 'providerNetworkError') return t('common.networkError');
+  if (message === 'providerReadResponseFailed' || message === 'providerClientInitFailed') return t('common.operationFailed');
+  if (message === 'providerRequestTimeout') return t('settings.providers.requestTimeout');
+  if (message.includes('API Key')) return t('settings.providers.apiKeyMissing');
+  if (message.includes('Endpoint')) return t('settings.providers.endpointMissing');
+  if (message.includes('timeout') || message.includes('超时')) return t('settings.providers.requestTimeout');
+  return t('common.operationFailed');
+}
+
+type CredentialFieldStatus = 'idle' | 'saving' | 'saved' | 'readError' | 'saveError' | 'copied' | 'copyError';
+
+interface CredentialFieldProps {
+  label: string;
+  account: string;
+  placeholder?: string;
+  mono?: boolean;
+  mask?: boolean;
+  defaultValue?: string;
+  trailing?: ReactNode;
+}
+
+function CredentialField({ label, account, placeholder, mono, mask, defaultValue, trailing }: CredentialFieldProps) {
+  const { t } = useTranslation();
+  const [value, setValue] = useState('');
+  const [revealed, setRevealed] = useState(false);
+  const [loaded, setLoaded] = useState(false);
+  const [dirty, setDirty] = useState(false);
+  const [status, setStatus] = useState<CredentialFieldStatus>('idle');
+  const debounceRef = useRef<number | null>(null);
+  const statusRef = useRef<number | null>(null);
+
+  useEffect(() => {
+    let cancelled = false;
+    setLoaded(false);
+    setDirty(false);
+    setStatus('idle');
+    setValue('');
+    if (debounceRef.current) {
+      clearTimeout(debounceRef.current);
+      debounceRef.current = null;
+    }
+    readCredential(account)
+      .then(v => {
+        if (cancelled) return;
+        setValue(v ?? '');
+        setLoaded(true);
+      })
+      .catch(error => {
+        if (cancelled) return;
+        console.error('[settings] failed to read credential', account, error);
+        setLoaded(true);
+        setStatus('readError');
+      });
+    return () => {
+      cancelled = true;
+    };
+  }, [account]);
+
+  useEffect(() => {
+    return () => {
+      if (debounceRef.current) clearTimeout(debounceRef.current);
+      if (statusRef.current) clearTimeout(statusRef.current);
+    };
+  }, []);
+
+  // 改造：除 readError（持续错误，留在输入旁标识字段不可用）外，所有 saving / saved /
+  //   saveError / copied / copyError 一律发到右上角 SavedToast。原内联文案太挤、跟其它
+  //   页面 toast 风格不统一。
+  const showTemporaryStatus = (next: CredentialFieldStatus) => {
+    if (next === 'saving') {
+      emitSaved('saving', t('common.saving'));
+    } else if (next === 'saved') {
+      emitSaved('saved', t('common.saved'));
+    } else if (next === 'saveError') {
+      emitSaved('failed', t('common.operationFailed'));
+    } else if (next === 'copied') {
+      emitSaved('saved', t('common.copied'));
+    } else if (next === 'copyError') {
+      emitSaved('failed', t('common.operationFailed'));
+    }
+    setStatus(next);
+    if (statusRef.current) clearTimeout(statusRef.current);
+    statusRef.current = window.setTimeout(() => setStatus('idle'), 1600);
+  };
+
+  const save = async (v: string, force = false) => {
+    if (!loaded || (!dirty && !force)) return;
+    setStatus('saving');
+    emitSaved('saving', t('common.saving'));
+    try {
+      await setCredential(account, v);
+      setDirty(false);
+      showTemporaryStatus('saved');
+    } catch (error) {
+      console.error('[settings] failed to save credential', account, error);
+      showTemporaryStatus('saveError');
+    }
+  };
+
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const v = e.target.value;
+    setValue(v);
+    if (!loaded) return;
+    setDirty(true);
+    if (debounceRef.current) clearTimeout(debounceRef.current);
+    debounceRef.current = window.setTimeout(() => save(v, true), 300);
+  };
+
+  const onBlur = () => {
+    if (!loaded || !dirty) return;
+    if (debounceRef.current) {
+      clearTimeout(debounceRef.current);
+      debounceRef.current = null;
+    }
+    save(value, true);
+  };
+
+  const fillDefault = async () => {
+    if (!loaded || !defaultValue) return;
+    setValue(defaultValue);
+    setDirty(true);
+    await save(defaultValue, true);
+  };
+
+  const onCopy = async () => {
+    if (!value || !loaded) return;
+    try {
+      if (!navigator.clipboard?.writeText) {
+        throw new Error('Clipboard API unavailable');
+      }
+      await navigator.clipboard.writeText(value);
+      showTemporaryStatus('copied');
+    } catch (error) {
+      console.error('[settings] failed to copy credential', account, error);
+      showTemporaryStatus('copyError');
+    }
+  };
+
+  const inputType = mask && !revealed ? 'password' : 'text';
+  const disabled = !loaded;
+
+  return (
+    <SettingRow label={label}>
+      <div style={{ display: 'flex', gap: 6, alignItems: 'center', width: '100%', maxWidth: 420 }}>
+        <input
+          type={inputType}
+          value={value}
+          placeholder={loaded ? placeholder : t('common.loading')}
+          onChange={handleChange}
+          onBlur={onBlur}
+          disabled={disabled}
+          style={{ ...inputStyle, fontFamily: mono ? 'var(--ol-font-mono)' : 'inherit' }}
+        />
+        {defaultValue && !value && loaded && (
+          <button onClick={fillDefault} title={t('settings.providers.fillDefault')} style={iconBtnStyle} disabled={!loaded}>
+            <Icon name="check" size={13} />
+          </button>
+        )}
+        {trailing}
+        {mask && (
+          <button
+            onClick={() => setRevealed(r => !r)}
+            title={revealed ? t('common.hide') : t('common.show')}
+            style={iconBtnStyle}
+            disabled={disabled}
+          >
+            <Icon name="eye" size={14} />
+          </button>
+        )}
+        <button
+          onClick={onCopy}
+          title={t('common.copy')}
+          style={iconBtnStyle}
+          disabled={!value || disabled}
+        >
+          <Icon name="copy" size={14} />
+        </button>
+        {/* readError 是字段无法读取的持续错误，留在原位提示用户该字段不可用；
+            其它瞬态状态（saving / saved / saveError / copied / copyError）都通过
+            emitSaved 发到右上角统一 toast，不再内联占位。 */}
+        {status === 'readError' && (
+          <span
+            style={{
+              fontSize: 11,
+              color: 'var(--ol-warn)',
+              whiteSpace: 'nowrap',
+            }}
+          >
+            {t('settings.providers.readFailed')}
+          </span>
+        )}
+      </div>
+    </SettingRow>
+  );
+}
+
+const miniBtnStyle: CSSProperties = {
+  height: 32, padding: '0 12px',
+  border: '0.5px solid var(--ol-line-strong)',
+  borderRadius: 8, background: 'var(--ol-surface)',
+  boxShadow: '0 1px 2px rgba(0,0,0,0.04), 0 0 0 0.5px rgba(255,255,255,0.2) inset',
+  color: 'var(--ol-ink-2)', cursor: 'default', flexShrink: 0,
+  fontSize: 12.5, fontWeight: 500, letterSpacing: '0.01em',
+  transition: 'background 0.16s var(--ol-motion-quick), border-color 0.16s var(--ol-motion-quick), color 0.16s var(--ol-motion-quick), box-shadow 0.16s var(--ol-motion-quick)',
+};
+
+const iconBtnStyle: CSSProperties = {
+  width: 32, height: 32,
+  border: '0.5px solid var(--ol-line-strong)',
+  borderRadius: 8, background: 'var(--ol-surface)',
+  boxShadow: '0 1px 2px rgba(0,0,0,0.04), 0 0 0 0.5px rgba(255,255,255,0.2) inset',
+  display: 'inline-flex', alignItems: 'center', justifyContent: 'center',
+  color: 'var(--ol-ink-3)', cursor: 'default', flexShrink: 0,
+  transition: 'background 0.16s var(--ol-motion-quick), border-color 0.16s var(--ol-motion-quick), color 0.16s var(--ol-motion-quick), transform 0.12s var(--ol-motion-quick)',
+};
