+++ openless-all/app/src/pages/settings/tabs.tsx
@@ -1,0 +1,78 @@
+// 设置弹窗里每个侧栏 tab 对应的内容页。每个 tab 就是若干 section 卡片的纵向堆叠；
+// 真正的逻辑都在各 *Section 文件里，这里只负责"哪些 section 归到哪个 tab"。
+
+import { useTranslation } from 'react-i18next';
+import { RecordingInputSection } from './RecordingInputSection';
+import { ShortcutsSection } from './ShortcutsSection';
+import { LanguageSection } from './LanguageSection';
+import { ProvidersSection } from './ProvidersSection';
+import { MarketplaceSection } from './MarketplaceSection';
+import { PermissionsSection } from './PermissionsSection';
+import { DataStorageSection } from './DataStorageSection';
+import { LocalModelSection } from './LocalModelSection';
+import { DebugToolsSection } from './DebugToolsSection';
+import { BetaChannelSection } from './BetaChannelSection';
+
+// 通用：录音与输入 · 快捷键 · 语言。
+export function GeneralTab() {
+  return (
+    <>
+      <RecordingInputSection />
+      <ShortcutsSection />
+      <LanguageSection />
+    </>
+  );
+}
+
+// 服务：AI 提供商 · 扩展市场。
+export function ServicesTab() {
+  return (
+    <>
+      <ProvidersSection />
+      <MarketplaceSection />
+    </>
+  );
+}
+
+// 隐私：本地优先说明 + 权限管理 · 数据存储。
+export function PrivacyTab() {
+  const { t } = useTranslation();
+  return (
+    <>
+      <div
+        style={{
+          display: 'flex',
+          alignItems: 'center',
+          gap: 10,
+          padding: '10px 12px',
+          borderRadius: 10,
+          background: 'var(--ol-blue-soft)',
+          marginBottom: 2,
+        }}
+      >
+        <span style={{
+          fontSize: 11, padding: '3px 8px', borderRadius: 999,
+          background: '#fff', color: 'var(--ol-blue)', fontWeight: 600, flexShrink: 0,
+        }}>
+          {t('modal.about.localFirst')}
+        </span>
+        <span style={{ fontSize: 11.5, color: 'var(--ol-ink-3)', lineHeight: 1.55 }}>
+          {t('modal.about.privacyDesc')}
+        </span>
+      </div>
+      <PermissionsSection />
+      <DataStorageSection />
+    </>
+  );
+}
+
+// 高级：本地模型 · 调试工具 · 加入 Beta 渠道（固定在最下面）。
+export function AdvancedTab() {
+  return (
+    <>
+      <LocalModelSection />
+      <DebugToolsSection />
+      <BetaChannelSection />
+    </>
+  );
+}
