+++ openless-all/app/src/styles/global.css
@@ -77,14 +77,33 @@ a { color: inherit; text-decoration: none; }
   }
 }
 
+/* 弹窗动画统一入口 —— 纯 opacity / transform，合成器友好。
+   背景不再逐帧动画 backdrop-filter（会卡顿、闪烁）；模糊由宿主元素静态设定，
+   动画只负责淡入淡出。设置弹窗 / 各市场弹窗 / GitHub 登录弹窗共用这套。 */
 @keyframes ol-modal-backdrop-in {
-  from { opacity: 0; backdrop-filter: blur(0); -webkit-backdrop-filter: blur(0); }
-  to   { opacity: 1; backdrop-filter: blur(8px) saturate(140%); -webkit-backdrop-filter: blur(8px) saturate(140%); }
+  from { opacity: 0; }
+  to   { opacity: 1; }
 }
 
 @keyframes ol-modal-backdrop-out {
-  from { opacity: 1; backdrop-filter: blur(8px) saturate(140%); -webkit-backdrop-filter: blur(8px) saturate(140%); }
-  to   { opacity: 0; backdrop-filter: blur(0); -webkit-backdrop-filter: blur(0); }
+  from { opacity: 1; }
+  to   { opacity: 0; }
+}
+
+@keyframes ol-modal-card-in {
+  from { opacity: 0; transform: translateY(8px) scale(0.985); }
+  to   { opacity: 1; transform: translateY(0) scale(1); }
+}
+
+/* 设置弹窗切换 tab 时右侧内容的轻微淡入。 */
+@keyframes ol-tab-fade {
+  from { opacity: 0; transform: translateY(4px); }
+  to   { opacity: 1; transform: translateY(0); }
+}
+
+/* 通用旋转 —— 「检查更新」按钮检查中的转圈动画等。 */
+@keyframes ol-spin {
+  to { transform: rotate(360deg); }
 }
 
 @keyframes ol-modal-drawer-in {
